Tests/LibWeb/Text/expected/UIEvents/KeyEvent-non-ascii.txt

keydown à
keydown á
keydown â
keydown ã
keydown ä
keydown å
